microsoft-git: update SHA-256 of latest installer (#27731)
After our automated creation of the manifest for microsoft-git version
2.33.0.vfs.0.0, we realized that the signing of the installer was not
quite correct. It was generated by a new automated process that did not
have all of the details correct. We replaced the installer on the GitHub
releases page with a new version that was generated from the same source
but used our older, manual process.

What we forgot to check was that the winget package stores the SHA-256
hash of the older installer, so "winget install microsoft-git" now fails
with a checksum mismatch.

I initially tried to re-run our process for generating the winget
manifest, but it failed to create the new commit, likely because a
manifest already existed. It did compute the latest SHA-256 for me [1].

[1] https://github.com/microsoft/git/runs/3614127425?check_suite_focus=true#step:2:87

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>