GnuWin32.Tree version 1.5.2.2 (#37973)
* GnuWin32.Tree version 1.5.2.2

* Update "PackageName" for Control Panel

The package name should be the same as in Control Panel

Co-authored-by: Levvie - she/her <11600822+ItzLevvie@users.noreply.github.com>

* Update "Publisher" for Control Panel

The Publisher information should be the same as in Control Panel.

Co-authored-by: Levvie - she/her <11600822+ItzLevvie@users.noreply.github.com>

* Update "MinimumOSVersion"

Windows is using "4-part version code". It should be complete.

Co-authored-by: Levvie - she/her <11600822+ItzLevvie@users.noreply.github.com>

* Update "InstallerUrl"

The previous URL for installer is copied from the offical homepage of "Tree for Windows". It should be changed to use the "direct link" from SourceForge.

Co-authored-by: Levvie - she/her <11600822+ItzLevvie@users.noreply.github.com>

Co-authored-by: Levvie - she/her <11600822+ItzLevvie@users.noreply.github.com>