Python.Python version 3.11.0 (#86190)
* Python.Python version 3.11.0

* Split changes to Python 3.10's moniker out to a separate PR

* Change UpgradeBehavior to uninstallPrevious

* Fix ProductCode to match GUID of uninstaller dir in Package Cache

My initial PR used the GUID for the Python 3.11.0 Core Interpreter, but
that is just one component of the total Python install, and just
uninstalling that will not remove Python but instead leave it in a bad
state. Updated to use the guid of the dir the uninstaller is stored in
inside `%LOCALAPPDATA%\Package Cache` instead.

* Add the ARM64 installer too