Total commander (#88222)
* Update Ghisler.TotalCommander.locale.de-DE.yaml

Changed ReleaseNoresURL to a more generic one to ease later updates. -> Only DownloadURLs and version numbers need to be changed in the future.

* Update Ghisler.TotalCommander.locale.en-US.yaml

Changed ReleaseNoresURL to a more generic one to ease later updates. -> Only DownloadURLs and version numbers need to be changed in the future.

* Update Ghisler.TotalCommander.locale.de-DE.yaml

Changed ReleaseNotesURL to a more generic one to ease later updates. -> Only DownloadURLs and version numbers need to be changed in the future.

* Update Ghisler.TotalCommander.locale.en-US.yaml

Changed ReleaseNotesURL to a more generic one to ease later updates. -> Only DownloadURLs and version numbers need to be changed in the future.